move cache.replace -> function
